requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.78.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypandoc==1.15
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pyzmq==26.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
